simplify data access patterns
Mutation of shared state was previously done in both event_loop.rs and ui_thread.rs. It is not consolidated into ui_thread.rs

Furthermore, platform::win32::window::WindowState is now read-only from WindowImpl, and write-only from ui_thread. This does cause a few issues by forcing RenderState to duplicate some data, but it greatly reduces the amount of thinking that needs to be done. Unfortunately, it does mean that all sorts of inputs have to be set twice, but WindowState is simple enough that this shouldn't be too great of a burden. We'll see, either way.