Add a global vsync thread and return to 2 threads per window
The 2 threads per window was done in order to see if having an event loop per thread would allow two windows that have been snapped side-by-side from the same process to both be repainted. This has turned out to not be the case. The event loop threads could be merged into one, though at the cost of more state management.